@@ -33,6 +33,7 @@ from metomi.isodatetime.dumpers import TimePointDumper
 from metomi.isodatetime.parsers import TimePointParser, DurationParser
 from metomi.isodatetime.exceptions import IsodatetimeError, ISO8601SyntaxError
 
+from cylc.flow import LOG
 from cylc.flow.parsec.exceptions import (
     ListValueError, IllegalValueError, IllegalItemError)
 from cylc.flow.subprocctx import SubFuncContext
@@ -236,9 +237,19 @@ class ParsecValidator:
                                 str(i) for i in cfg[key] if i not in voptions
                             ]
                             if bad:
-                                raise IllegalValueError(
+                                exc = IllegalValueError(
                                     'option', [*keys, key], ', '.join(bad)
                                 )
+                                if specval.warn_options:
+                                    LOG.warning(
+                                        f'{exc}'
+                                        '\nInvalid items have been removed'
+                                    )
+                                    cfg[key] = [
+                                        x for x in cfg[key] if x not in bad
+                                    ]
+                                else:
+                                    raise exc
                         elif cfg[key] not in voptions:
                             raise IllegalValueError(
                                 'option', [*keys, key], cfg[key]
