@@ -34,10 +34,10 @@ from subprocess import (
 import sys
 from time import sleep
 from typing import (
-    TYPE_CHECKING,
     Any,
     Dict,
     List,
+    Literal,
     Optional,
     Tuple,
     Union,
@@ -55,10 +55,6 @@ from cylc.flow.platforms import (
     get_platform,
 )
 from cylc.flow.util import format_cmd
-
-
-if TYPE_CHECKING:
-    from typing_extensions import Literal
 
 
 def get_proc_ancestors():
@@ -83,8 +79,10 @@ def get_proc_ancestors():
 def watch_and_kill(proc):
     """Kill proc if my PPID (etc.) changed - e.g. ssh connection dropped."""
     gpa = get_proc_ancestors()
+    # Allow customising the interval to allow tests to run faster:
+    interval = float(os.getenv('CYLC_PROC_POLL_INTERVAL', 60))
     while True:
-        sleep(0.5)
+        sleep(interval)
         if proc.poll() is not None:
             break
         if get_proc_ancestors() != gpa:
