@@ -48,8 +48,9 @@ from typing import (
 import cylc.flow.flags
 from cylc.flow import LOG
 from cylc.flow.async_util import make_async
+from cylc.flow.cfgspec.glbl_cfg import glbl_cfg
 from cylc.flow.exceptions import (
-    ContactFileExists,
+    SchedulerAlive,
     CylcError,
     InputError,
     ServiceFileError,
@@ -372,6 +373,7 @@ def _is_process_running(
 ) -> bool:
     """Check if a workflow process is still running.
 
+    Runs `cylc psutil` on the scheduler run host.
     * Returns True if the process is still running.
     * Returns False if it is not.
     * Raises CylcError if we cannot tell (e.g. due to network issues).
@@ -428,8 +430,13 @@ def _is_process_running(
         text=True
     )  # * hardcoded command
     try:
-        # Terminate command after 10 seconds to prevent hanging, etc.
-        out, err = proc.communicate(timeout=10, input=metric)
+        out, err = proc.communicate(
+            timeout=(
+                glbl_cfg().get(
+                    ['scheduler', 'run hosts', 'process check timeout'])
+            ),
+            input=metric
+        )
     except TimeoutExpired:
         raise CylcError(
             f'Attempt to determine whether workflow is running on {host}'
@@ -479,7 +486,7 @@ def detect_old_contact_file(
     * If one does exist but the workflow process is definitely not alive,
       remove it.
     * If one exists and the workflow process is still alive, raise
-      ContactFileExists.
+      SchedulerAlive.
 
     Args:
         id_: workflow ID
@@ -489,7 +496,7 @@ def detect_old_contact_file(
         CylcError:
             If it is not possible to tell for sure if the workflow is running
             or not.
-        ContactFileExists:
+        SchedulerAlive:
             If old contact file exists and the workflow process still alive.
         ServiceFileError:
             For corrupt / incompatible contact files.
@@ -522,7 +529,7 @@ def detect_old_contact_file(
     fname = get_contact_file_path(id_)
     if process_is_running:
         # ... the process is running, raise an exception
-        raise ContactFileExists(
+        raise SchedulerAlive(
             CONTACT_FILE_EXISTS_MSG % {
                 "host": old_host,
                 "port": old_port,
