@@ -21,6 +21,7 @@ from typing import (
     List,
     cast,
 )
+from unittest.mock import Mock
 
 import pytest
 
@@ -53,6 +54,7 @@ from cylc.flow.task_outputs import (
 from cylc.flow.task_state import (
     TASK_STATUS_FAILED,
     TASK_STATUS_PREPARING,
+    TASK_STATUS_RUNNING,
     TASK_STATUS_SUCCEEDED,
     TASK_STATUS_WAITING,
 )
@@ -124,7 +126,7 @@ def get_pb_prereqs(schd: 'Scheduler') -> 'List[PbPrerequisite]':
 
 
 @pytest.fixture(scope='module')
-async def harness(mod_flow, mod_scheduler, mod_start):
+async def mod_harness(mod_flow, mod_scheduler, mod_start):
     flow_def = {
         'scheduler': {
             'allow implicit tasks': True
@@ -206,22 +208,22 @@ def collect_states(data, node_type):
     ]
 
 
-def test_generate_definition_elements(harness):
+def test_generate_definition_elements(mod_harness):
     """Test method that generates all definition elements."""
-    schd, data = harness
+    schd, data = mod_harness
     task_defs = schd.config.taskdefs.keys()
     assert len(data[TASKS]) == len(task_defs)
     assert len(data[TASK_PROXIES]) == len(task_defs)
 
 
-def test_generate_graph_elements(harness):
-    schd, data = harness
+def test_generate_graph_elements(mod_harness):
+    schd, data = mod_harness
     task_defs = schd.config.taskdefs.keys()
     assert len(data[TASK_PROXIES]) == len(task_defs)
 
 
-def test_get_data_elements(harness):
-    schd, data = harness
+def test_get_data_elements(mod_harness):
+    schd, data = mod_harness
     flow_msg = schd.data_store_mgr.get_data_elements(TASK_PROXIES)
     assert len(flow_msg.added) == len(data[TASK_PROXIES])
 
@@ -232,16 +234,16 @@ def test_get_data_elements(harness):
     assert len(none_msg.ListFields()) == 0
 
 
-def test_get_entire_workflow(harness):
+def test_get_entire_workflow(mod_harness):
     """Test method that populates the entire workflow protobuf message."""
-    schd, data = harness
+    schd, data = mod_harness
     flow_msg = schd.data_store_mgr.get_entire_workflow()
     assert len(flow_msg.task_proxies) == len(data[TASK_PROXIES])
 
 
-def test_increment_graph_window(harness):
+def test_increment_graph_window(mod_harness):
     """Test method that adds and removes elements window boundary."""
-    schd, data = harness
+    schd, data = mod_harness
     assert schd.data_store_mgr.prune_trigger_nodes
     assert len(data[TASK_PROXIES]) == 2
 
@@ -253,10 +255,10 @@ def test_in_window_extra_edges(edgeharness):
     assert f'{w_id}//$edge|1/c1|1/c2' in data[EDGES]
 
 
-def test_initiate_data_model(harness):
+def test_initiate_data_model(mod_harness):
     """Test method that generates all data elements in order."""
     schd: Scheduler
-    schd, data = harness
+    schd, data = mod_harness
     assert len(data[WORKFLOW].task_proxies) == 2
     schd.data_store_mgr.initiate_data_model(reloaded=True)
     assert len(data[WORKFLOW].task_proxies) == 2
@@ -268,10 +270,10 @@ def test_initiate_data_model(harness):
     assert schd.data_store_mgr.n_edge_distance == 2
 
 
-async def test_delta_task_state(harness):
+async def test_delta_task_state(mod_harness):
     """Test update_data_structure. This method will generate and
     apply adeltas/updates given."""
-    schd, data = harness
+    schd, data = mod_harness
     # follow only needs to happen once .. tests working on the same object?
     w_id = schd.data_store_mgr.workflow_id
     schd.data_store_mgr.data[w_id] = data
@@ -289,11 +291,11 @@ async def test_delta_task_state(harness):
     await schd.update_data_structure()
 
 
-async def test_delta_task_held(harness):
+async def test_delta_task_held(mod_harness):
     """Test update_data_structure. This method will generate and
     apply adeltas/updates given."""
     schd: Scheduler
-    schd, data = harness
+    schd, data = mod_harness
     schd.pool.hold_tasks(['*'])
     await schd.update_data_structure()
     assert True in {t.is_held for t in data[TASK_PROXIES].values()}
@@ -312,27 +314,31 @@ async def test_delta_task_held(harness):
     await schd.update_data_structure()
 
 
-def test_insert_job(harness):
+def test_insert_job(mod_harness):
     """Test method that adds a new job to the store."""
-    schd, data = harness
+    schd: Scheduler
+    schd, data = mod_harness
     assert len(schd.data_store_mgr.added[JOBS]) == 0
-    schd.data_store_mgr.insert_job('foo', '1', 'submitted', job_config(schd))
+    itask = schd.pool.get_tasks()[0]
+    schd.data_store_mgr.insert_job(itask, 'submitted', job_config(schd))
     assert len(schd.data_store_mgr.added[JOBS]) == 1
     assert ext_id(schd) in schd.data_store_mgr.added[JOBS]
 
 
-def test_insert_db_job(harness, job_db_row):
+def test_insert_db_job(mod_harness, job_db_row):
     """Test method that adds a new job from the db to the store."""
-    schd, data = harness
+    schd: Scheduler
+    schd, data = mod_harness
     assert len(schd.data_store_mgr.added[JOBS]) == 1
     schd.data_store_mgr.insert_db_job(0, job_db_row)
     assert len(schd.data_store_mgr.added[JOBS]) == 2
     assert ext_id(schd) in schd.data_store_mgr.added[JOBS]
 
 
-def test_delta_job_msg(harness):
+def test_delta_job_msg(mod_harness):
     """Test method adding messages to job element."""
-    schd, data = harness
+    schd: Scheduler
+    schd, data = mod_harness
     j_id = ext_id(schd)
     tokens = Tokens(j_id)
     # First update creation
@@ -341,22 +347,26 @@ def test_delta_job_msg(harness):
     assert schd.data_store_mgr.updated[JOBS][j_id].messages
 
 
-def test_delta_job_attr(harness):
+def test_delta_job_attr(mod_harness):
     """Test method modifying job fields to job element."""
-    schd, data = harness
+    schd: Scheduler
+    schd, data = mod_harness
     schd.data_store_mgr.delta_job_attr(
-        Tokens(ext_id(schd)), 'job_runner_name', 'at')
+        Mock(job_tokens=Tokens(ext_id(schd))), 'job_runner_name', 'at'
+    )
     assert schd.data_store_mgr.updated[JOBS][ext_id(schd)].messages != (
         schd.data_store_mgr.added[JOBS][ext_id(schd)].job_runner_name
     )
 
 
-def test_delta_job_time(harness):
+def test_delta_job_time(mod_harness):
     """Test method setting job state change time."""
-    schd, data = harness
+    schd: Scheduler
+    schd, data = mod_harness
     event_time = get_current_time_string()
     schd.data_store_mgr.delta_job_time(
-        Tokens(ext_id(schd)), 'submitted', event_time)
+        Mock(job_tokens=Tokens(ext_id(schd))), 'submitted', event_time
+    )
     job_updated = schd.data_store_mgr.updated[JOBS][ext_id(schd)]
     with pytest.raises(ValueError):
         job_updated.HasField('jumped_time')
@@ -365,10 +375,10 @@ def test_delta_job_time(harness):
     )
 
 
-async def test_update_data_structure(harness):
+async def test_update_data_structure(mod_harness):
     """Test update_data_structure. This method will generate and
     apply adeltas/updates given."""
-    schd, data = harness
+    schd, data = mod_harness
     w_id = schd.data_store_mgr.workflow_id
     schd.data_store_mgr.data[w_id] = data
     schd.pool.hold_tasks(['*'])
@@ -487,10 +497,10 @@ async def test_prune_data_store(flow, scheduler, start):
         }) == 0
 
 
-async def test_family_ascent_point_prune(harness):
+async def test_family_ascent_point_prune(mod_harness):
     """Test _family_ascent_point_prune. This method tries to remove
     non-existent family."""
-    schd, data = harness
+    schd, data = mod_harness
     fp_id = 'NotAFamilyProxy'
     parent_ids = {fp_id}
     checked_ids = set()
@@ -506,10 +516,10 @@ async def test_family_ascent_point_prune(harness):
     assert len(parent_ids) == 0
 
 
-def test_delta_task_prerequisite(harness):
+def test_delta_task_prerequisite(mod_harness):
     """Test delta_task_prerequisites."""
     schd: Scheduler
-    schd, data = harness
+    schd, data = mod_harness
     schd.pool.set_prereqs_and_outputs(
         schd.pool.get_task_ids(),
         [(TASK_STATUS_SUCCEEDED,)],
@@ -718,7 +728,7 @@ async def test_delta_task_outputs(one: 'Scheduler', start):
         assert itask
         itask.submit_num += 1
         one.data_store_mgr.insert_job(
-            itask.tdef.name, itask.point, itask.state.status, {'submit_num': 1}
+            itask, itask.state.status, {'submit_num': 1}
         )
         await one.update_data_structure()
 
@@ -814,3 +824,24 @@ async def test_log_events(one: Scheduler, start):
             assert log_record.message == 'here hare here'
         finally:
             LOG.removeHandler(handler)
+
+
+async def test_no_backwards_job_state_change(one: Scheduler, start):
+    """It should not allow backwards job state changes."""
+    def get_job_state(itask):
+        return one.data_store_mgr.data[one.id][JOBS][itask.job_tokens.id].state
+
+    async with start(one):
+        itask = one.pool.get_tasks()[0]
+        itask.state_reset(TASK_STATUS_PREPARING)
+        itask.submit_num += 1
+        await one.update_data_structure()
+
+        one.task_events_mgr.process_message(itask, INFO, TASK_OUTPUT_STARTED)
+        await one.update_data_structure()
+        assert get_job_state(itask) == TASK_STATUS_RUNNING
+
+        # Simulate late arrival of "submitted" message
+        one.task_events_mgr.process_message(itask, INFO, TASK_OUTPUT_SUBMITTED)
+        await one.update_data_structure()
+        assert get_job_state(itask) == TASK_STATUS_RUNNING
