@@ -14,12 +14,18 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+import asyncio
 import logging
+import sys
 from typing import (
     Any as Fixture,
     Callable
 )
 
+if sys.version_info[:2] >= (3, 11):
+    from asyncio import timeout as async_timeout
+else:
+    from async_timeout import timeout as async_timeout
 import pytest
 
 from cylc.flow.commands import (
@@ -31,7 +37,10 @@ from cylc.flow.commands import (
     set_prereqs_and_outputs,
 )
 from cylc.flow.cycling.integer import IntegerPoint
-from cylc.flow.task_state import TASK_STATUS_WAITING
+from cylc.flow.task_state import (
+    TASK_STATUS_WAITING,
+    TASK_STATUS_RUNNING
+)
 
 
 async def test_trigger_workflow_paused(
@@ -617,3 +626,129 @@ async def test_trigger_n0_tasks(
             # downstream task
             ('z', '[1, 2, 3, 4, 5]'),
         }
+
+
+async def test_replay_outputs(flow, scheduler, start, complete, log_filter):
+    """Triggered group start tasks re-emit (kind of) earlier outputs.
+
+    https://github.com/cylc/cylc-flow/issues/6858
+
+    Example graph:
+        a:started => b => end
+        k:kustom => l => end
+        k:kustom => offg
+
+    If I trigger a, b, k, l AFTER a:started and k:kustom have completed:
+        cylc trigger workflow //1/a //1/b //1/k //1/l
+
+    Then I should expect outputs `k:kustom` and `a:started` to be re-used
+    to satify b and l in the triggered flow, but NOT off-group task offg.
+    """
+    msg_prereq = '[1/{}:waiting(runahead)] prerequisite force-satisfied: 1/{}'
+    msg_spawned = "[1/{}:waiting(runahead)] => waiting"
+    msg_removed = "Removed tasks: 1/{}"
+
+    wid = flow({
+        'scheduling': {
+            'graph': {
+                'R1': """
+                    a:started => b => end
+                    k:kustom => l => end
+                    k:kustom => offg
+                """
+            }
+        },
+        'runtime': {
+            'a': {},
+            'k': {
+                'outputs': {'kustom': 'custom message'}
+            }
+        }
+    })
+    schd = scheduler(wid, paused_start=True)
+    async with start(schd):
+        # Set initial tasks a and k to "running" so they are recognized as
+        # live during the forthcoming trigger operation.
+
+        # Complete the a:started and k:kustom outputs.
+        schd.pool.set_prereqs_and_outputs(
+            ['1/a'], ['started'], [], []
+        )
+        schd.pool.set_prereqs_and_outputs(
+            ['1/k'], ['kustom'], [], []
+        )
+        # It should spawn b, l, and offg.
+        for task in ['b', 'l', 'offg']:
+            assert log_filter(contains=msg_spawned.format(task))
+
+        # Set a and k as running so they're recognized as live start tasks
+        # by the trigger operation.
+        for itask in schd.pool.get_tasks():
+            itask.state_reset(TASK_STATUS_RUNNING)
+
+        # Now trigger the group.
+        await run_cmd(
+            force_trigger_tasks(schd, ['1/a', '1/b', '1/k', '1/l'], [])
+        )
+        # It should remove b and l (in-group tasks)
+        for task in ['b', 'l']:
+            assert log_filter(contains=msg_removed.format(task))
+
+        # But they will be respawned immediately by re-satisfying dependence
+        # on the earlier outputs of a and k for in-group tasks:
+        assert log_filter(contains=msg_prereq.format('b', 'a:started'))
+        assert log_filter(contains=msg_prereq.format('l', 'k:custom message'))
+        # But not for the off-group task offg:
+        assert not log_filter(
+            contains=msg_prereq.format('offg', 'k:custom message'))
+
+
+async def test_trigger_with_sequential_task(flow, scheduler, run, log_filter):
+    """It should trigger a failed sequential task.
+
+    See https://github.com/cylc/cylc-flow/issues/6911
+    """
+    id_ = flow({
+        'scheduling': {
+            'initial cycle point': '1',
+            'final cycle point': '2',
+            'cycling mode': 'integer',
+            'special tasks': {
+                'sequential': 'foo',
+            },
+            'graph': {
+                'R1': 'install => foo',
+                'P1': 'foo',
+            },
+        },
+        'runtime': {
+            'foo': {
+                'simulation': {
+                    'fail cycle points': '2',
+                },
+            },
+        },
+    })
+    schd = scheduler(id_, paused_start=False)
+    async with run(schd):
+        # wait for 2/foo:failed
+        async with async_timeout(5):
+            while True:
+                itask = schd.pool._get_task_by_id('2/foo')
+                if itask and itask.state.outputs.is_message_complete('failed'):
+                    break
+                await asyncio.sleep(0)
+
+        # re-trigger 2/foo
+        await run_cmd(
+            force_trigger_tasks(
+                schd, ['2/foo'], []
+            )
+        )
+
+        # it should re-run
+        async with async_timeout(5):
+            while True:
+                if log_filter(contains='[2/foo/02:running] (received)failed'):
+                    break
+                await asyncio.sleep(0)
