@@ -15,19 +15,12 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import asyncio
-import sys
 from types import MethodType
 
 import pytest
 
 from cylc.flow.scheduler import SchedulerError
 from cylc.flow.workflow_events import WorkflowEventHandler
-
-
-if sys.version_info[:2] >= (3, 11):
-    from asyncio import timeout as async_timeout
-else:
-    from async_timeout import timeout as async_timeout
 
 
 EVENTS = (
@@ -51,20 +44,23 @@ async def test_scheduler(flow, scheduler, capcall):
     def get_events():
         return {e[0][1] for e in events}
 
-    def _schd(config=None, **opts):
-        id_ = flow({
-            'scheduler': {
-                'events': {
-                    'mail events': ', '.join(EVENTS),
-                    **(config or {}),
+    def _schd(event_config=None, config=None, **opts):
+        assert not (event_config and config)
+        if not config:
+            config = {
+                'scheduler': {
+                    'events': {
+                        'mail events': ', '.join(EVENTS),
+                        **(event_config or {}),
+                    },
                 },
-            },
-            'scheduling': {
-                'graph': {
-                    'R1': 'a'
-                }
-            },
-        })
+                'scheduling': {
+                    'graph': {
+                        'R1': 'a'
+                    }
+                },
+            }
+        id_ = flow(config)
         schd = scheduler(id_, **opts)
         schd.get_events = get_events
         return schd
@@ -92,7 +88,7 @@ async def test_workflow_timeout(test_scheduler, run):
     This counts down from scheduler start.
     """
     schd = test_scheduler({'workflow timeout': 'PT0S'})
-    async with async_timeout(4):
+    async with asyncio.timeout(4):
         async with run(schd):
             await asyncio.sleep(0.1)
     assert schd.get_events() == {'startup', 'workflow timeout', 'shutdown'}
@@ -107,7 +103,7 @@ async def test_inactivity_timeout(test_scheduler, start):
         'inactivity timeout': 'PT0S',
         'abort on inactivity timeout': 'True',
     })
-    async with async_timeout(4):
+    async with asyncio.timeout(4):
         with pytest.raises(SchedulerError):
             async with start(schd):
                 await asyncio.sleep(0)
@@ -168,12 +164,57 @@ async def test_stall(test_scheduler, start):
     assert schd.get_events() == {'shutdown', 'stall'}
 
 
-async def test_restart_timeout(test_scheduler, scheduler, run, complete):
-    """Test restart timeout.
+async def test_restart_timeout_workflow_completion(
+    test_scheduler,
+    scheduler,
+    run,
+    complete,
+):
+    """Test restart timeout for completed workflows.
 
     This should fire when a completed workflow is restarted.
     """
     schd = test_scheduler({'restart timeout': 'PT0S'}, paused_start=False)
+
+    # run to completion
+    async with run(schd):
+        await complete(schd)
+    assert schd.get_events() == {'startup', 'shutdown'}
+
+    # restart
+    schd2 = scheduler(schd.workflow)
+    schd2.get_events = schd.get_events
+    async with run(schd2):
+        await asyncio.sleep(0.1)
+    assert schd2.get_events() == {'startup', 'restart timeout', 'shutdown'}
+
+
+async def test_restart_timeout_workflow_stop_after_cycle_point(
+    test_scheduler,
+    scheduler,
+    run,
+    complete,
+):
+    """Test restart timeout with the "stop after cycle point" config.
+
+    This should fire when a completed workflow is restarted.
+    """
+    schd = test_scheduler(
+        config={
+            'scheduler': {
+                'cycle point format': 'CCYY',
+                'events': {'restart timeout': 'PT0S'},
+            },
+            'scheduling': {
+                'initial cycle point': '2000',
+                'stop after cycle point': '2000',
+                'graph': {
+                    'P1Y': 'foo[-P1Y] => foo',
+                },
+            },
+        },
+        paused_start=False,
+    )
 
     # run to completion
     async with run(schd):
@@ -220,7 +261,7 @@ async def test_shutdown_handler_timeout_kill(
         '''
     )
 
-    async with async_timeout(30):
+    async with asyncio.timeout(30):
         async with run(schd):
             # Replace a scheduler method, to call handlers in simulation mode.
             monkeypatch.setattr(
